Reorganised AdminPortal pages.
