+++ front-end-203/src/App.js
@@ -5,12 +5,12 @@ import HomePage from "./homepage/HomePage";
 import FlightSearch from "./flight-search/FlightSearch";
 import SignIn from "./signin-screen/SignInPage";
 import SignUp from "./signup-screen/SignUpPage";
-import EditProfile from "./adminPortal/EditProfile";
-import PlaneAdd from "./adminPortal/PlaneAdd";
+import EditProfile from "./admin-portal/EditProfile";
+import PlaneAdd from "./admin-portal/AdminManagePlane";
 
 // Admin Portal
-import AdminPortalLogin from "./adminPortal/AdminPortalLogin";
-import AdminPortalHomePage from "./adminPortal/AdminPortalHomePage";
+import AdminPortalLogin from "./admin-portal/AdminPortalLogin";
+import AdminPortalHomePage from "./admin-portal/AdminPortalHomePage";
 
 function App() {
   return (
@@ -20,12 +20,14 @@ function App() {
         <Route path="/flightsearch" element={<FlightSearch />} />
         <Route path="/signin" element={<SignIn />} />
         <Route path="/signup" element={<SignUp />} />
+
+        {/* TODO: This should be moved out of admin pages and then made generic when you click on the profile icon on the header bar */}
         <Route path="/editprofile" element={<EditProfile />} />
-        <Route path="/planeadd" element={<PlaneAdd />} />
 
         {/* ADMIN PORTAL */}
         <Route path="/adminPortal/home" element={<AdminPortalHomePage />} />
         <Route path="/adminPortal/login" element={<AdminPortalLogin />} />
+        <Route path="/adminPortal/planes" element={<PlaneAdd />} />
       </Routes>
     </Router>
   );
